requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-redis==4.10.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
redis==3.3.0
sqlparse==0.3.0
